+++ osbrain/tests/test_agent_async_requests_handlers.py
@@ -8,7 +8,6 @@ from osbrain import run_agent
 from osbrain.helper import wait_agent_attr
 
 from .common import append_received
-from .common import nsproxy  # noqa: F401
 
 
 class ServerAsyncRep(Agent):
