+++ osbrain/tests/test_agent_pubsub_topics.py
@@ -6,7 +6,6 @@ from osbrain.helper import last_received_endswith
 from osbrain.helper import wait_agent_condition
 
 from .common import append_received
-from .common import nsproxy  # noqa: F401
 
 
 @pytest.mark.parametrize(
