+++ osbrain/tests/test_agent_req_rep.py
@@ -8,8 +8,6 @@ from osbrain import run_logger
 from osbrain.helper import logger_received
 from osbrain.helper import sync_agent_logger
 
-from .common import nsproxy  # noqa: F401
-
 
 def test_return(nsproxy):
     """
