+++ osbrain/tests/test_agent_sync_publications.py
@@ -14,7 +14,6 @@ from osbrain.helper import wait_agent_attr
 from osbrain.helper import wait_agent_condition
 
 from .common import append_received
-from .common import nsproxy  # noqa: F401
 
 
 class BaseServer(Agent):
