+++ osbrain/tests/test_bugs.py
@@ -6,8 +6,6 @@ import sys
 from osbrain import run_agent
 from osbrain.helper import wait_agent_condition
 
-from .common import nsproxy  # noqa: F401
-
 
 def test_timer_recursion(nsproxy):
     """
