+++ osbrain/tests/test_helper.py
@@ -16,9 +16,6 @@ from osbrain.helper import wait_agent_attr
 from osbrain.helper import wait_agent_condition
 from osbrain.helper import wait_condition
 
-from .common import agent_logger  # noqa: F401
-from .common import nsproxy  # noqa: F401
-
 
 class SugarAddict():
     def __init__(self):
