+++ osbrain/tests/test_timer.py
@@ -12,7 +12,6 @@ from osbrain.common import repeat
 from osbrain.helper import wait_agent_attr
 
 from .common import append_received
-from .common import nsproxy  # noqa: F401
 from .common import set_received
